Update README.md
PDF forensic tool is aimed to extract deep information from a PDF file which is not visible generally. If you ever receive an anonymous PDF file from somewhere, or from an unknown source, it will be possible to find out the Author of the file, when and how it's created and modified.